third_party/WebKit/LayoutTests/accessibility/aom-computed-accessible-node.html

<!DOCTYPE HTML>
<script src="../resources/gc.js"></script>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<!--

Accessibility Object Model
Explainer: https://github.com/WICG/aom/blob/master/explainer.md
Spec: https://wicg.github.io/aom/spec/

-->

<p id="button1" role="button">Click</p>

<script>

test(function(t) {
    assert_true(internals.runtimeFlags.accessibilityObjectModelEnabled);
}, "Make sure that Accessibility Object Model is enabled");

promise_test(() => {
    var element = document.getElementById("button1");
    return window.getComputedAccessibleNode(element).then(computedAXNode => {
      assert_true(computedAXNode != null);
    });
}, 'Ensure that a non null value is returned from getComputedAccessibleNode');

promise_test(() => {
    var element = document.getElementById("button1");
    return window.getComputedAccessibleNode(element).then(computedAXNode => {
      assert_equals(computedAXNode.role, "button");
    });
}, "ComputedAccessibleNode.role");

promise_test(() => {
    var element = document.getElementById("button1");
    return window.getComputedAccessibleNode(element).then(computedAXNode => {
      assert_equals(computedAXNode.name, "Click");
    });
}, "ComputedAccessibleNode.name");

</script>

<p id="button2" role="button" aria-label="axButton">Click</p>

<script>

promise_test(() => {
    var element = document.getElementById("button2");
    return window.getComputedAccessibleNode(element).then(computedAXNode => {
      assert_equals(computedAXNode.name, "axButton");
    });
}, "ComputedAccessibleNode.name set through ARIA.");

promise_test(() => {
    var element = document.getElementById("button2");
    return window.getComputedAccessibleNode(element).then(computedAXNode1 => {
      window.getComputedAccessibleNode(element).then(computedAXNode2 => {
        assert_true(computedAXNode1 === computedAXNode2);
      });
    });
}, "Multiple calls to getComputedAccessibleNode will return the same node.");


promise_test(() => {
    var button1 = document.getElementById("button1");
    var button2 = document.getElementById("button2");
    return window.getComputedAccessibleNode(button2).then(button2AXNode => {
      assert_equals(button2AXNode.name, "axButton");
      assert_equals(button2AXNode.role, "button");

      // Set to display:none to remove from accessibility tree.
      button2.style.display = 'none';
      window.getComputedAccessibleNode(button1).then(button1AXNode => {
        // button1's computedAXNode should still be valid.
        assert_equals(button1AXNode.name, "Click");
        assert_equals(button1AXNode.role, "button");

        // button2s computedAXNode should only contain empty attributes.
        assert_equals(button2AXNode.name, "");
        assert_equals(button2AXNode.role, "");
      })
    });

}, "Deleting nodes from the accessibility tree will not cause a crash, and set computed properties on any references to a deleted computed accessible node to empty.");

</script>
